refactor(orchestrator): collapse server/client orchestrators onto one impl
createPluginOrchestrator.{server,client}.ts had near-identical ~60-line bodies
differing only in which dev-server / react-static plugin they pull (.server vs
.client) and the transformer's defaultEnvironment. Extract the shared body to
createPluginOrchestrator.impl.ts (createPluginOrchestratorImpl + an
OrchestratorStrategy of {defaultEnvironment, devServerPlugin, staticPlugin}); each
side shrinks to a ~10-line binding.

The server-first variant gated the SSG plugin on an always-true capability while
the client-first pushed it unconditionally — identical behavior, so the impl
pushes it unconditionally.

Left untouched: each file's exported  interface. It is NOT the same as
orchestrator/types.ts's Strategy (this one has , that one
), and plugin/types.ts imports this one — consolidating the
two Strategy types is a separate concern, out of scope here.

Verified both conditions: build clean, 101 entrypoints, test:unit 472,
test:server 655/4-skip, test:client 182/7-skip. Closes 66g.6.
